refactor: use `type` for types
