+++ packages/generators/src/types/index.ts
@@ -1,6 +1,5 @@
-import Generator from "yeoman-generator";
-import path from "path";
-import { IWebpackCLI } from "webpack-cli";
+import type Generator from "yeoman-generator";
+import { type IWebpackCLI } from "webpack-cli";
 
 export type InitOptions = { template: string; force?: boolean };
 export type LoaderOptions = { template: string };
@@ -20,32 +19,3 @@ export type CustomGeneratorOptions<T extends BaseCustomGeneratorOptions> =
     cli: IWebpackCLI;
     options: T;
   };
-
-export class CustomGenerator<
-  T extends BaseCustomGeneratorOptions = BaseCustomGeneratorOptions,
-  Z extends CustomGeneratorOptions<T> = CustomGeneratorOptions<T>,
-> extends Generator<Z> {
-  public cli: IWebpackCLI;
-  public template: string;
-  public dependencies: string[];
-  public force: boolean;
-  public answers: Record<string, unknown>;
-  public generationPath: string;
-  public supportedTemplates: string[];
-  public packageManager: string | undefined;
-
-  public constructor(args: string | string[], opts: Z) {
-    super(args, opts);
-
-    this.cli = opts.cli;
-    this.dependencies = [];
-    this.answers = {};
-    this.supportedTemplates = [];
-
-    const { options } = opts;
-
-    this.template = options.template;
-    this.force = typeof options.force !== "undefined" ? options.force : false;
-    this.generationPath = path.resolve(process.cwd(), options.generationPath);
-  }
-}
