+++ packages/generators/src/utils/helpers.ts
@@ -1,4 +1,4 @@
-import { CustomGenerator } from "../types";
+import type { CustomGenerator } from "../custom-generator";
 import { List } from "./scaffold-utils";
 
 const regex = /[A-Z]{2,}(?=[A-Z][a-z]+[0-9]*|\b)|[A-Z]?[a-z]+[0-9]*|[A-Z]|[0-9]+/g;
