+++ packages/webpack-cli/src/index.ts
@@ -1,5 +1,5 @@
-import { IWebpackCLI } from "./types";
-export * from "./types";
+import { type IWebpackCLI } from "./types";
+export type * from "./types";
 
 // eslint-disable-next-line @typescript-eslint/no-var-requires
 const CLI: IWebpackCLI = require("./webpack-cli");
