+++ packages/webpack-cli/src/types.ts
@@ -1,4 +1,4 @@
-import webpack, {
+import type {
   EntryOptions,
   Stats,
   Configuration,
@@ -12,16 +12,17 @@ import webpack, {
   AssetEmittedInfo,
   FileCacheOptions,
 } from "webpack";
+import type webpack from "webpack";
 
 // eslint-disable-next-line @typescript-eslint/ban-ts-comment
 // @ts-ignore extraneous import is intended
 // eslint-disable-next-line node/no-extraneous-import
-import { ClientConfiguration, Configuration as DevServerConfig } from "webpack-dev-server";
+import type { ClientConfiguration, Configuration as DevServerConfig } from "webpack-dev-server";
 
-import { Colorette } from "colorette";
-import { Command, CommandOptions, Option, ParseOptions } from "commander";
-import { prepare } from "rechoir";
-import { stringifyStream } from "@discoveryjs/json-ext";
+import { type Colorette } from "colorette";
+import { type Command, type CommandOptions, type Option, type ParseOptions } from "commander";
+import { type prepare } from "rechoir";
+import { type stringifyStream } from "@discoveryjs/json-ext";
 
 /**
  * Webpack CLI
